requirements.txt
streamlit
pandas
request
scikit-learn
